Refactor: Update attendance data handling to support holiday filtering and improve working days calculation

+++ app/src/main/java/com/example/sheetcompute/domain/excel/ExcelRowParser.kt
@@ -47,7 +47,7 @@ object ExcelRowParser {
         val date = parseDateSafely(dateStr) ?: return null
         val clockIn = calcClockIn(timeCell)?: return null
         val tardy = if (clockIn.isAfter(workStartTime)) {
-            Duration.between(workStartTime, clockIn).toMinutes().toInt()
+            Duration.between(workStartTime, clockIn).toMinutes()
         } else 0
         val record = AttendanceRecord(
             employeeId = id.toString(),
